Adding better handling of NULL terminated strings to parsing code.
I noticed that sometimes an item called "amul01" (example) would not be present in the index although they key "amul01_<type>" was indeed present. This was because "amul01" was a string loaded from an item description file and it had a \u0000 character at the end.

This commit tries to addres those cases by removing NULL chars at the end of these strings. Providing better interaction between different modules of the app.